+++ CHANGELOG.md
@@ -25,9 +25,17 @@ Note that the displayed date is in the format `dd-mm-yyyy`
 
 - Change `modLogChan` to `modChannel` in db
 
+## [v1.10.3]
+
+> **Released:** `09-07-2021`
+
+### Bug fixes
+
+- Fix test command
+
 ## [v1.10.2]
 
-> **Released:** `TBA`
+> **Released:** `06-07-2021`
 
 ### Bug fixes
 
@@ -356,6 +364,7 @@ Note that the displayed date is in the format `dd-mm-yyyy`
 [#8]: https://github.com/Welcome-Bot/welcome-bot/pull/8
 [#3]: https://github.com/Welcome-Bot/welcome-bot/pull/3
 [#6]: https://github.com/Welcome-Bot/welcome-bot/pull/6
+[v1.10.3]: https://github.com/Welcome-Bot/welcome-bot/releases/tag/v1.10.3
 [v1.10.2]: https://github.com/Welcome-Bot/welcome-bot/releases/tag/v1.10.2
 [v1.10.1]: https://github.com/Welcome-Bot/welcome-bot/releases/tag/v1.10.1
 [v1.10.0]: https://github.com/Welcome-Bot/welcome-bot/releases/tag/v1.10.0